Create data.json
Name is up for debate, but essentially, this just stores important data for the application itself, not so much for the editor or output. Things like first run flags, update information, etc. This will be an ongoing list, but I think it's important to maintain some kind of human-readable settings.